Quarkus lambda client - addressed PR comments
