+++ lambda/deployment/src/test/java/io/quarkus/amazon/lambdaclient/deployment/LambdaSyncClientFullConfigTest.java
