+++ lambda/deployment/src/test/resources/sync-urlconn-full-config.properties
@@ -1,0 +1,10 @@
+quarkus.lambda.endpoint-override=http://localhost:9090
+
+quarkus.lambda.aws.region=us-east-2
+quarkus.lambda.aws.credentials.type=static
+quarkus.lambda.aws.credentials.static-provider.access-key-id=test-key
+quarkus.lambda.aws.credentials.static-provider.secret-access-key=test-secret
+
+quarkus.lambda.sync-client.type = url
+quarkus.lambda.sync-client.connection-timeout = 0.100S
+quarkus.lambda.sync-client.socket-timeout = 0.100S
