+++ lambda/runtime/src/main/java/io/quarkus/amazon/lambda/client/runtime/LambdaBuildTimeConfig.java
