+++ lambda/runtime/src/main/java/io/quarkus/amazon/lambda/client/runtime/LambdaDevServicesBuildTimeConfig.java
