--- lambdaclient/pom.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <parent>
-        <groupId>io.quarkiverse.amazonservices</groupId>
-        <artifactId>quarkus-amazon-services-build-parent</artifactId>
-        <version>999-SNAPSHOT</version>
-        <relativePath>../build-parent/pom.xml</relativePath>
-    </parent>
-
-    <artifactId>quarkus-amazon-lambda-client-parent</artifactId>
-    <name>Quarkus - Amazon Services - Lambda client</name>
-    <packaging>pom</packaging>
-
-    <modules>
-        <module>runtime</module>
-        <module>deployment</module>
-    </modules>
-
-
-</project>
